Relocate NetworksInventory to nucypher.blockchain.eth.networks
